refactor(gulp-paeckchen): move host to own module
