+++ packages/gulp-paeckchen/test/host-test.ts
@@ -1,0 +1,14 @@
+import test from 'ava';
+import { File } from 'gulp-util';
+
+import { GulpHost } from '../src/host';
+
+test('GulpHost should return previously added files', t => {
+  const host = new GulpHost();
+  const file = new File({
+    path: '/test'
+  });
+
+  host.addFile(file);
+  t.is(host.getFile(file.path), file);
+});
